Separate declaration and implementation in ttkUtils
